requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
django-allauth==0.58.2
python-dotenv==1.0.0
django-cleanup==8.0.0
django-widget-tweaks==1.5.0
